Pulling image if not present
In case that docker daemon returns 404 it may mean that user provided
image by name, and that image isn't pulled. We are than trying to
interpret image id as repo:tag combo and pull it. If that succeeds,
We are using id of newly pulled image and trying to create container
again.